backend/requirements.txt
Django==3.2.3
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
drf-spectacular==0.24.2
gunicorn==20.1.0    
Pillow==9.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
webcolors==1.11.1
